Update and clean a version of the final report
I’m reasonably confident that there was a bug before the ramp rate
calculation in the final report. Checking the change in generation from
previous hour for the data used to calculate ramp rate showed lots of
nan’s and the values didn’t match what had been previously calculated.

All of these files and changes are in preparation for the video
recording.